+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Moved `extract_patch` to start session request to avoid starting a patch extractor side-car container if not needed. **Breaking change.**
+- `extract_patch` parameter now forces the archive to be a git repository. **Breaking change.**
+- Migrated project from Python 3.12 to Python 3.14. **Breaking change.**
+- Updated dependencies.
+
+### Fixed
+
+- Fixed `.daiv.yml` schema to the newer version of `daiv`.
+
 ## [0.2.0] - 2025-08-31
 
 ### Added